requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
supabase>=2.0.0
lightgbm>=4.1.0
scikit-learn>=1.3.0

